feat: 게시판, 댓글 crud 구현완료, swagger 적용

+++ src/main/java/com/ddokdoghotdog/gowalk/post/dto/response/PostGetDetailResponseDTO.java
@@ -25,7 +25,7 @@ public class PostGetDetailResponseDTO {
     private Long memberid;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private int likesCount;
+    private Long likesCount;
     private int commentsCount;
     private List<String> imgList;
     private List<CommentDTO> comments;
